# Massively optimise hydro and dragon egg gens