--- a/src/main/java/me/profelements/dynatech/items/electric/generators/ChippingGenerator.java
+++ b/src/main/java/me/profelements/dynatech/items/electric/generators/ChippingGenerator.java
@@ -1,10 +1,5 @@
 package me.profelements.dynatech.items.electric.generators;
 
-import org.bukkit.Location;
-import org.bukkit.Material;
-import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.meta.Damageable;
-
 import me.mrCookieSlime.CSCoreLibPlugin.Configuration.Config;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
@@ -12,51 +7,56 @@ import me.mrCookieSlime.Slimefun.api.BlockStorage;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
 import me.profelements.dynatech.items.electric.abstracts.AMachineGenerator;
+import org.bukkit.Location;
+import org.bukkit.Material;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.Damageable;
+import org.bukkit.inventory.meta.ItemMeta;
 
 import javax.annotation.Nonnull;
 
 public class ChippingGenerator extends AMachineGenerator {
 
-    private static final int PowerPerDurability = 8;
+    private final int powerPerDurability = 8;
+
+    private final ItemStack progressBar = new ItemStack(Material.WOODEN_AXE);
 
     public ChippingGenerator(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe);
     }
 
-    
+
     @Override
     public int getGeneratedOutput(Location l, Config data) {
         BlockMenu inv = BlockStorage.getInventory(l.getBlock());
-        int julesAmount = 0;
+        int julesAmount;
 
         for (int slot : getInputSlots()) {
             ItemStack item = inv.getItemInSlot(slot);
-            if (item != null && item.getItemMeta() instanceof Damageable ) {
-                Damageable im = (Damageable) item.getItemMeta();
-                if (!im.hasDamage()) {
-                    
-                    julesAmount = item.getType().getMaxDurability()*PowerPerDurability;
+            // Do as many lightweight checks as possible before we do the intensive stuff
+            if (item != null && !item.getType().isAir() && item.getType().isItem() && item.hasItemMeta()) {
+                // `getItemMeta` does multiple clones! Even doing this once is slow, nevermind multiple times!
+                ItemMeta meta = item.getItemMeta();
+                if (meta instanceof Damageable) {
+                    Damageable im = (Damageable) meta;
+                    if (!im.hasDamage()) {
+                        julesAmount = item.getType().getMaxDurability() * powerPerDurability;
+                        if (julesAmount != 0) {
+                            inv.consumeItem(slot);
+                            return julesAmount;
+                        }
+                    }
+
+                    julesAmount = (item.getType().getMaxDurability() - im.getDamage()) * 2;
                     if (julesAmount != 0) {
                         inv.consumeItem(slot);
                         return julesAmount;
                     }
-                    
                 }
-
-                julesAmount = (item.getType().getMaxDurability()- im.getDamage())*2;
-                if (julesAmount != 0) {
-                    inv.consumeItem(slot);
-                    return julesAmount;
-                }
-                
             }
         }
         return 0;
-
-
-
     }
-
 
     @Override
     public String getMachineIdentifier() {
@@ -66,7 +66,6 @@ public class ChippingGenerator extends AMachineGenerator {
     @Nonnull
     @Override
     public ItemStack getProgressBar() {
-        return new ItemStack(Material.WOODEN_AXE);
+        return progressBar;
     }
-    
 }