--- a/src/main/java/me/profelements/dynatech/items/electric/generators/DragonEggGenerator.java
+++ b/src/main/java/me/profelements/dynatech/items/electric/generators/DragonEggGenerator.java
@@ -1,20 +1,16 @@
 package me.profelements.dynatech.items.electric.generators;
 
+import io.github.thebusybiscuit.slimefun4.core.attributes.EnergyNetProvider;
 import me.mrCookieSlime.CSCoreLibPlugin.Configuration.Config;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
-import me.mrCookieSlime.Slimefun.api.BlockStorage;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
-import me.profelements.dynatech.DynaTech;
-
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.inventory.ItemStack;
-
-import io.github.thebusybiscuit.slimefun4.core.attributes.EnergyNetProvider;
 
 import javax.annotation.Nonnull;
 
@@ -26,29 +22,17 @@ public class DragonEggGenerator extends SlimefunItem implements EnergyNetProvide
 
     @Override
     public int getGeneratedOutput(@Nonnull Location location, @Nonnull Config config) {
-        if (DynaTech.getInstance().getTickInterval()  % 10 == 0) {
-            
-            Block dragonEgg = location.getBlock().getRelative(BlockFace.UP);
-            if (dragonEgg.getType() == Material.DRAGON_EGG) {
-                BlockStorage.addBlockInfo(location, "egg", String.valueOf(true));
-                return getEnergyProduction();
-            } else {
-                BlockStorage.addBlockInfo(location, "egg", String.valueOf(false));
-            }
-
-         } else if (BlockStorage.getLocationInfo(location, "egg") != null && BlockStorage.getLocationInfo(location, "egg").equals(String.valueOf(true))){
-            return getEnergyProduction();
+        Block dragonEgg = location.getBlock().getRelative(BlockFace.UP);
+        if (dragonEgg.getType() == Material.DRAGON_EGG) {
+            return 32;
         }
-        return 0; 
+
+        return 0;
     }
 
     @Override
     public boolean isChargeable() {
         return false;
-    }
-
-    public static final int getEnergyProduction() {
-        return 32;
     }
 
     @Override