--- a/src/main/java/me/profelements/dynatech/items/misc/DimensionalHomeDimension.java
+++ b/src/main/java/me/profelements/dynatech/items/misc/DimensionalHomeDimension.java
@@ -1,13 +1,12 @@
 package me.profelements.dynatech.items.misc;
 
-import java.util.Random;
-
 import org.bukkit.Material;
 import org.bukkit.World;
 import org.bukkit.generator.ChunkGenerator;
 
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
+import java.util.Random;
 
 public class DimensionalHomeDimension extends ChunkGenerator {
 
@@ -19,22 +18,22 @@ public class DimensionalHomeDimension extends ChunkGenerator {
 
         chunkData.setRegion(0, 59, 0, 16, 60, 16, Material.BEDROCK);
         for (int y = 60; y < 180; y++) {
-            for (int x = 0; x <16; x++) {
+            for (int x = 0; x < 16; x++) {
                 chunkData.setBlock(x, y, 0, Material.BARRIER);
                 chunkData.setBlock(x, y, 16, Material.BARRIER);
             }
-            for (int z = 0; z <16; z++) {
+            for (int z = 0; z < 16; z++) {
                 chunkData.setBlock(0, y, z, Material.BARRIER);
                 chunkData.setBlock(16, y, z, Material.BARRIER);
             }
-            
+
         }
-        for(int x2 = 0; x2 < 16; x2++) {
+        for (int x2 = 0; x2 < 16; x2++) {
             for (int y2 = 0; y2 < 16; y2++) {
                 chunkData.setBlock(x2, 180, y2, Material.BARRIER);
             }
         }
         return chunkData;
-    } 
+    }
 
 }