There have been multiple improvements in this PR:
* Massively improved hydro gens
  - Use a Guava Cache to cache the energy values. These will be refreshed every 1 min (ish).
  - Remove the BlockStorage stuff which is SLOW - BlockStorage serializes and deserializes a ton which is slow.
* Optimised Chipping Generator
  - getProgressBar was returning a NEW object everytime, fixed that
  - Avoid any expensive item checks (with != null, != air, etc)
  - Avoid multiple calls to getItemMeta - getItemMeta will do **multiple** clones, it's slow! Make sure to only ever do this once
* Optimised dragon egg gen
  - The things you did to try and optimise this just slow it down. While creating a new Block instance isn't great it's a hell of a lot quicker than a lot of BlockStorage stuff
* Removed some redundant runSync's
* Fixed general formatting issues and inconsistencies 

Hydro gen before:
![image](https://user-images.githubusercontent.com/8492901/120902198-d17b8980-c636-11eb-83d3-4b635d13aa39.png)

After:
```
[19:30:49 INFO]: Slimefun 4> Please wait a second... The results are coming in!
[19:30:49 INFO]: 
[19:30:49 INFO]: ===== Slimefun Lag Profiler =====
[19:30:49 INFO]: Total time: 0.55ms
[19:30:49 INFO]: Running every: 0.6s (12 ticks)
[19:30:49 INFO]: Performance: :::::::::::::::::::: - Good (1%)
[19:30:49 INFO]: 
[19:30:49 INFO]: 32 blocks
  ...
  WATER_MILL - 3x (0.01ms | avg: 0ms)
  WATER_TURBINE - 3x (0ms | avg: 0ms)
```

---
**Note**: I have not tested this extensively. I have only tested the hydro stuff some. Theoretically it should all be fine regardless but yeah