@WalshyDev 2021-06-06
That's what happens when you change something at like midnight on phone :p

I didn't realise you were using CSLib's Config which doesn't support default